requirements.txt
sentence-transformers
google-generativeai
python-dotenv
pymongo
flask